refactor(controller): refactor controller package
1. refactor controller package structure
