+++ src/main/java/j2b/nft_generator/test/TestController.java
@@ -11,24 +11,5 @@ import org.springframework.web.bind.annotation.GetMapping;
 @Controller
 public class TestController {
 
-    @GetMapping("/item")
-    public String item(Model model) {
-        return "item";
-    }
-
-    @GetMapping("/addItem")
-    public String addItem(Model model) {
-        return "addItemForm";
-    }
-
-    @GetMapping("/loginForm")
-    public String login(Model model) {
-        return "login";
-    }
-
-    @GetMapping("/signupForm")
-    public String signUp(Model model) {
-        return "signup";
-    }
 
 }
